+++ app/sensor_series.py
@@ -2,18 +2,31 @@ import plotly.graph_objects as go
 
 
 class SensorSeriesVisualisation(go.Figure):
-    def __init__(self, sensors_data):
+    def __init__(self, traces, plot_title=""):
         super().__init__()
-        for data in sensors_data:
+        for trace_name, color in traces:
             self.add_trace(go.Scatter(
-                name=data["name"],
+                name=trace_name,
                 mode='markers+lines',
-                x=data["measurement_date"],
-                y=data["sensor_values"],
-                text=[f"Value: {sensor_value}<br>Anomaly: {anomaly}<br>Trace: {trace_name}"
-                      for sensor_value, anomaly, trace_name
-                      in zip(data["sensor_values"], data["anomalies"], data["trace_name"])],
+                marker_color=color,
+                x=[],
+                y=[],
+                text=[],
                 hovertemplate="%{text}<extra></extra>",
             ))
 
-        self.update_layout(template="plotly_white", hovermode="x unified", xaxis_title="Date", yaxis_title="Value")
+        self.update_layout(template="plotly_white", hovermode="x unified", xaxis_title="Date", yaxis_title="Value",
+                           title={"text": f"<b>{plot_title}</b>"})
+
+    @staticmethod
+    def update_figure_data(measurements_data, figure, foot="L"):
+        for i in range(0, 3):
+            trace_value_key = f"{foot}{i}_value"
+            figure["data"][i]["x"] = measurements_data["measurement_date"]
+            figure["data"][i]["y"] = measurements_data[trace_value_key]
+            figure["data"][i]["text"] = [
+                f"Value: {value}<br>Anomaly: {anomaly}<br>Trace: {measurements_data['trace_name']}"
+                for value, anomaly in
+                zip(measurements_data[trace_value_key], measurements_data[f"{foot}{i}_anomaly"])]
+
+        return figure
