+++ docker-compose.yml
@@ -6,7 +6,7 @@ services:
             context: ./app
         environment:
             - MONGO_URL=mongodb://mongo_db:27017
-#            - DATA_SOURCE_URL=http://server:3000/tesla/
+#            - DATA_SOURCE_URL=http://0.0.0.0:3000/tesla/
             - DATA_SOURCE_URL=http://tesla.iem.pw.edu.pl:9080/v2/monitor/
             - ISOD_LOGIN=${ISOD_LOGIN}
             - ISOD_PASSWORD=${ISOD_PASSWORD}
